refactor: more optimizing
