+++ frontend/src/hooks/useAllocationAgreement.js
@@ -1,7 +1,6 @@
 import { apiRoutes } from '@/constants/routes'
 import { useApiService } from '@/services/useApiService'
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
-import { useCurrentUser } from './useCurrentUser'
 
 export const useAllocationAgreementOptions = (params, options) => {
   const client = useApiService()
@@ -76,22 +75,6 @@ export const useSaveAllocationAgreement = (params, options) => {
         'compliance-report-summary',
         params.complianceReportId
       ])
-    }
-  })
-}
-
-export const useGetAllocationAgreementsChangeLog = ({
-  complianceReportGroupUuid
-}) => {
-  const client = useApiService()
-  const path = apiRoutes.getChangelog
-    .replace(':complianceReportGroupUuid', complianceReportGroupUuid)
-    .replace(':dataType', 'allocation-agreements')
-  return useQuery({
-    queryKey: ['allocation-agreements-changelog', complianceReportGroupUuid],
-    queryFn: async () => {
-      const response = await client.get(path)
-      return response.data
     }
   })
 }
