+++ frontend/src/hooks/useFuelExport.js
@@ -1,7 +1,6 @@
 import { apiRoutes } from '@/constants/routes'
 import { useApiService } from '@/services/useApiService'
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
-import { useCurrentUser } from './useCurrentUser'
 
 export const useFuelExportOptions = (params, options) => {
   const client = useApiService()
@@ -69,20 +68,6 @@ export const useSaveFuelExport = (params, options) => {
         'compliance-report-summary',
         params.complianceReportId
       ])
-    }
-  })
-}
-
-export const useGetFuelExportsChangeLog = ({ complianceReportGroupUuid }) => {
-  const client = useApiService()
-  const path = apiRoutes.getChangelog
-    .replace(':complianceReportGroupUuid', complianceReportGroupUuid)
-    .replace(':dataType', 'fuel-exports')
-  return useQuery({
-    queryKey: ['fuel-exports-log', complianceReportGroupUuid],
-    queryFn: async () => {
-      const response = await client.get(path)
-      return response.data
     }
   })
 }
