+++ frontend/src/views/FuelExports/FuelExportChangelog.jsx
@@ -1,9 +1,11 @@
 import { BCGridViewer } from '@/components/BCDataGrid/BCGridViewer'
 import BCTypography from '@/components/BCTypography'
 import Loading from '@/components/Loading'
-import { useGetComplianceReport } from '@/hooks/useComplianceReports'
+import {
+  useGetChangeLog,
+  useGetComplianceReport
+} from '@/hooks/useComplianceReports'
 import { useCurrentUser } from '@/hooks/useCurrentUser'
-import { useGetFuelExportsChangeLog } from '@/hooks/useFuelExport'
 import colors from '@/themes/base/colors'
 import { Box } from '@mui/material'
 import { useTranslation } from 'react-i18next'
@@ -23,9 +25,10 @@ export const FuelExportChangelog = () => {
   )
 
   const { data: changelogData, isLoading: changelogDataLoading } =
-    useGetFuelExportsChangeLog({
+    useGetChangeLog({
       complianceReportGroupUuid:
-        currentReportData.report.complianceReportGroupUuid
+        currentReportData.report.complianceReportGroupUuid,
+      dataType: 'fuel-exports'
     })
 
   const getRowId = (params) => {
